+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-alpha.2
+
+* Limits the scope of application of the method while the filter is applied.
+  - insert, insertAll, removeAt.
+* Add removeLastFromOriginal.
+
 ## 0.0.1-alpha.1
 
 * Limits the scope of application of the method while the filter is applied.